fix bugs

+++ app/src/main/java/com/iti/mobile/triporganizer/data/room/dao/NoteDao.java
@@ -25,7 +25,7 @@ public interface NoteDao {
     void deleteNote(Note note);
 
     @Query("SELECT * FROM notes WHERE tripId = :tripId")
-    LiveData<List<Note>> getAllNote(int tripId);
+    LiveData<List<Note>> getAllNote(long tripId);
 
     @Query("SELECT * FROM notes WHERE tripId = :tripId")
     List<Note> getAllNoteNotLive(int tripId);
